+++ app/src/main/java/com/example/foodie_finder/ui/viewModel/ProfileViewModel.kt
@@ -3,7 +3,7 @@ package com.example.foodie_finder.ui.viewModel
 import android.graphics.Bitmap
 import androidx.lifecycle.ViewModel
 import com.example.foodie_finder.data.local.User
-import com.example.foodie_finder.data.local.UserModel
+import com.example.foodie_finder.data.model.UserModel
 
 class ProfileViewModel : ViewModel() {
